requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyautogui==0.9.54
openai==1.3.0
google-generativeai==0.3.1
Pillow==9.5.0
pytesseract==0.3.10
PyAudio==0.2.13
numpy==1.24.3
pyperclip==1.8.2
sounddevice==0.4.6
soundfile==0.12.1
mediapipe==0.10.0
pycaw==20220416
screen-brightness-control==0.9.0
python-dotenv==1.0.0
Flask==2.3.3
Flask-CORS==4.0.0